google_contacts_sync/requirements.txt
cachetools
certifi
charset-normalizer
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
oauthlib
protobuf
pyasn1
pyasn1-modules
pyparsing
requests
requests-oauthlib
rsa
six
uritemplate
urllib3
